Enable OMP toolset for the project Rewrite SurfaceExt::BlurRect

+++ src/Ext/Rules/Body.h
@@ -63,7 +63,7 @@ public:
 		Valueable<bool> IronCurtain_KeptOnDeploy;
 		Valueable<ColorStruct> ToolTip_Background_Color;
 		Valueable<int> ToolTip_Background_Opacity;
-		Valueable<int> ToolTip_Background_BlurSize;
+		Valueable<float> ToolTip_Background_BlurSize;
 
 		ExtData(RulesClass* OwnerObject) : Extension<RulesClass>(OwnerObject)
 			, Storage_TiberiumIndex { -1 }
@@ -98,7 +98,7 @@ public:
 			, IronCurtain_KeptOnDeploy { true }
 			, ToolTip_Background_Color { {0, 0, 0} }
 			, ToolTip_Background_Opacity { 100 }
-			, ToolTip_Background_BlurSize { 0 }
+			, ToolTip_Background_BlurSize { 0.f }
 		{ }
 
 		virtual ~ExtData() = default;
