+++ src/Ext/Surface/Body.h
@@ -5,6 +5,6 @@
 class SurfaceExt : public Surface
 {
 public:
-	void BlurRect(const RectangleStruct& rect, int blurSize);
+	void BlurRect(const RectangleStruct& rect, float blurSize);
 
 };
